additional validation on custom forms

+++ backend/tables/form_views.py
@@ -3,7 +3,7 @@ from rest_framework.decorators import api_view
 from rest_framework import status
 from backend.tables.models import *
 from backend.tables.serializers import *
-from backend.config.form_config import FORM_FIELDS
+from backend.config.form_config import *
 
 
 @api_view(['GET', 'POST'])
@@ -24,20 +24,18 @@ def form_data(request, model_pk):
         if not UserType.contains_user(request.user, "models"):
             return Response(
                 {"permission_error": ["User does not have permission."]}, status=status.HTTP_401_UNAUTHORIZED)
-        # TODO: check if has custom form cal mode
-        # TODO: check indexes and fieldtypes exist
-        # TODO: sort by index
-        # validate field types
-        for field in request.data['fields']:
-            field['itemmodel'] = model_pk
-            field['cal_event'] = None
-            if field['fieldtype'] not in FORM_FIELDS.values():
-                return Response({"form_error": ["Invalid field type."]}, status=status.HTTP_400_BAD_REQUEST)
-        # TODO: add additional validation?
+        if "custom_form" not in itemmodel.calibrationmode_set.values_list("name", flat=True):
+            return Response({"form_error": ["Item model does not allow custom forms."]}, status=status.HTTP_404_NOT_FOUND)
+        # validate fields
+        errors = validate_new_form(request.data['fields'], model_pk)
+        if len(errors) > 0:
+            return Response({"form_error": errors}, status=status.HTTP_400_BAD_REQUEST)
+
+        fields = sorted(request.data['fields'], key=lambda i: i['index'])
         # delete previous form if exists
         if len(itemmodel.calibrationformfield_set.all()):
             itemmodel.calibrationformfield_set.all().delete()
-        serializer = FormFieldSerializer(data=request.data['fields'], many=True)
+        serializer = FormFieldSerializer(data=fields, many=True)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -57,13 +55,20 @@ def submit_form(request, cal_event_pk):
             return Response(
                 {"permission_error": ["User does not have permission."]}, status=status.HTTP_401_UNAUTHORIZED)
 
+        itemmodel = cal_event.instrument.item_model
+        if "custom_form" not in itemmodel.calibrationmode_set.values_list("name", flat=True):
+            return Response({"form_error": ["Item model does not allow custom forms."]}, status=status.HTTP_400_BAD_REQUEST)
+        if len(itemmodel.calibrationformfield_set.all() == 0):
+            return Response({"form_error": ["Item model does not have a custom form."]}, status=status.HTTP_400_BAD_REQUEST)
+
         if len(cal_event.calibrationformfield_set.all()) > 0:
             return Response({"form_error": ["Calibration event already has form associated."]}, status=status.HTTP_400_BAD_REQUEST)
 
-        errors = validate_form(request.data['fields'], cal_event_pk)
+        errors = validate_form_submit(request.data['fields'], cal_event_pk)
         if len(errors) > 0:
             return Response({"form_error": errors}, status=status.HTTP_400_BAD_REQUEST)
-        serializer = FormFieldSerializer(data=request.data['fields'], many=True)
+        fields = sorted(request.data['fields'], key=lambda i: i['index'])
+        serializer = FormFieldSerializer(data=fields, many=True)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -79,25 +84,64 @@ def submit_form(request, cal_event_pk):
         return Response({'cal_event': cal_event_serializer.data, 'fields': form_serializer.data}, status=status.HTTP_200_OK)
 
 
-def validate_form(fields, cal_event_pk):
+def validate_new_form(fields, model_pk):
+    errors = []
+    for field in fields:
+        field['itemmodel'] = model_pk
+        field['cal_event'] = None
+        if 'index' not in field or 'fieldtype' not in field:
+            errors.append({'error': "Index and fieldtype are required for all fields."})
+            continue
+        if field['fieldtype'] not in FORM_FIELDS.values():
+            errors.append({'index': field['index'], 'error': "Invalid field type."})
+            continue
+
+        expected_fields = EXPECTED_FIELDS[field['fieldtype']]
+        unexpected_fields = ALL_FIELDS - expected_fields
+        missing = []
+        for name in expected_fields:
+            if name not in field or field[name] is None or field[name] == "":
+                missing.append(name)
+        for name in unexpected_fields:
+            if name in field: field[name] = None
+        if len(missing) > 0:
+            errors.append({'index': field['index'], 'error': "Missing required fields for this fieldtype.", 'missing': missing})
+            continue
+
+        if field['fieldtype'] == FORM_FIELDS['float_input'] and field['expected_min'] > field['expected_max']:
+            errors.append({'index': field['index'], 'error': "Min cannot be greater than max."})
+    return errors
+
+
+def validate_form_submit(fields, cal_event_pk):
     errors = []
     for field in fields:
         field['cal_event'] = cal_event_pk
         field['itemmodel'] = None
+        if 'index' not in field or 'fieldtype' not in field:
+            errors.append({'error': "Index and fieldtype are required for all fields."})
+            continue
 
         if field['fieldtype'] == FORM_FIELDS['header'] or field['fieldtype'] == FORM_FIELDS['plaintext'] or field['fieldtype'] == FORM_FIELDS['text_input']:
             field['value_okay'] = True
         elif field['fieldtype'] == FORM_FIELDS['bool_input']:
-            if not field['actual_bool']:
-                errors.append({'index': field['index'], 'error': "Value must be true."})
-            else:
-                field['value_okay'] = True
+            try:
+                actual_bool = field['actual_bool']
+                if not actual_bool:
+                    errors.append({'index': field['index'], 'error': "Value must be true."})
+                else:
+                    field['value_okay'] = True
+            except KeyError:
+                errors.append({'index': field['index'], 'error': "Actual boolean not provided."})
         elif field['fieldtype'] == FORM_FIELDS['float_input']:
-            val = field['actual_float']
-            if (field['expected_min'] and val < field['expected_min']) or (field['expected_max'] and val > field['expected_max']):
-                errors.append({'index': field['index'], 'error': "Value is outside acceptable range."})
-            else:
-                field['value_okay'] = True
+            try:
+                val = field['actual_float']
+                if (field['expected_min'] and val < field['expected_min']) or (field['expected_max'] and val > field['expected_max']):
+                    errors.append({'index': field['index'], 'error': "Value is outside acceptable range."})
+                else:
+                    field['value_okay'] = True
+            except KeyError:
+                errors.append({'index': field['index'], 'error': "Actual float not provided."})
         else:
             errors.append({'index': field['index'], 'error': "Invalid field type."})
     return errors
